R Markdown DfE National Statistic Output Format

This function creates an R Markdown output format that generates documents in the style of DfE National Statistic Publications.

national_statistic(...)

Arguments

...
All Parameters from rmarkdown::word_document are accepted

Details

To format an R Markdown document using this template add the function to the R Markdowns YAML header as follows:

title: "Name"
output: dfeR::national_statistic

You can also use the template in R Studio via File > New File > R Markdown > From Template ...

Examples

national_statistic()

#> $knitr
#> $knitr$opts_knit
#> NULL
#> 
#> $knitr$opts_chunk
#> $knitr$opts_chunk$dev
#> [1] "png"
#> 
#> $knitr$opts_chunk$dpi
#> [1] 96
#> 
#> $knitr$opts_chunk$fig.width
#> [1] 5
#> 
#> $knitr$opts_chunk$fig.height
#> [1] 4
#> 
#> 
#> $knitr$knit_hooks
#> NULL
#> 
#> $knitr$opts_hooks
#> NULL
#> 
#> $knitr$opts_template
#> NULL
#> 
#> 
#> $pandoc
#> $pandoc$to
#> [1] "docx"
#> 
#> $pandoc$from
#> [1] "markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart"
#> 
#> $pandoc$args
#> [1] "--highlight-style"                                                                  
#> [2] "tango"                                                                              
#> [3] "--reference-doc"                                                                    
#> [4] "/home/travis/build/dfe-analytical-services/dfeR/inst/styles/national_statistic.docx"
#> 
#> $pandoc$keep_tex
#> [1] FALSE
#> 
#> $pandoc$latex_engine
#> [1] "pdflatex"
#> 
#> $pandoc$ext
#> NULL
#> 
#> 
#> $keep_md
#> [1] FALSE
#> 
#> $clean_supporting
#> [1] TRUE
#> 
#> $df_print
#> [1] "default"
#> 
#> $pre_knit
#> NULL
#> 
#> $post_knit
#> NULL
#> 
#> $pre_processor
#> function (metadata, input_file, runtime, knit_meta, files_dir, 
#>     output_dir) 
#> {
#>     saved_files_dir <<- files_dir
#>     NULL
#> }
#> <environment: 0x4f08980>
#> 
#> $intermediates_generator
#> function (...) 
#> {
#>     reference_intermediates_generator(saved_files_dir, ..., reference_docx)
#> }
#> <environment: 0x4f08980>
#> 
#> $post_processor
#> NULL
#> 
#> $on_exit
#> NULL
#> 
#> attr(,"class")
#> [1] "rmarkdown_output_format"