Read and clean a .sql script for use with RODBC

This function allows you to read in and clean .sql string that can be used directly with RODBC.

It removes statements such as USE and GO, all -- comments and ensures formatting is correct. Without this lots of scripts fail when read in and used with RODBC.

read_sql_script(file)

Arguments

file
The .sql file to be read in

Value

Character vector of clean sql query

Examples

# NOT RUN {
read_sql_script("Queries/script_name.sql")
# }